Build in a pure Wayland environment
Note that this change only allow the package to build in a pure wayland
environment but doesn't make it functionnal since some X11 -> Wayland
equivalents doesn't exist yet. The changes that remain to be made in
order to get it working in a pure Wayland environment are marked by the
comment "TO_DO_WAYLAND". Bug report regarding this issue: PTREL-724.

Change-Id: I69194192d39605e9858141af7f1e74c4206f32c5
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>